# Speed up java-to-managed typemap lookups