… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Resources/Base/BuildReleaseArm64SimpleDotNet.apkdesc
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Resources/Base/BuildReleaseArm64SimpleLegacy.apkdesc
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Resources/Base/BuildReleaseArm64XFormsDotNet.apkdesc
… src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Resources/Base/BuildReleaseArm64XFormsLegacy.apkdesc
… src/monodroid/jni/embedded-assemblies.cc
… src/monodroid/jni/embedded-assemblies.hh
… src/monodroid/jni/monodroid-glue-internal.hh
… src/monodroid/jni/monodroid-glue.cc
… src/monodroid/jni/xamarin-app.hh