@grendello 2022-04-11
> This looks ok. I'd need to read up on the LLVM code bits in order to review those changes properly, but the rest looks ok.

If you want to know how they work, I suggest reading the code outside the diff and I'm glad to explain and answer any questions :)