@grendello 2022-04-11
> > If you want to know how they work, I suggest reading the code outside the diff and I'm glad to explain and answer any questions :)
> 
> OK, how does LLVM work? rofl

It's complicated :laughing: