@grendello 2022-04-11
Seriously though, what we (and, of course, other clang frontends - that is specific language compilers) generate is LLVM IR (Intermediate Representation) code which allows one to describe both data in a type-safe manner that is abstract enough to allow for different translations into native assemblers and also allows easy(ier) optimization of the **IR** code before it's translated into the target architecture code. The target code generator takes the already somewhat optimized IR code and generates the most optimal native assembler for the given abstract construct. That's a very, very rough approximation of how it works :)