@jonpryor 2022-04-11
The commit message states:

> * a simpler binary search function was implemented

I think this could use some elaboration. :-)  What makes it simpler?  Why change it at all?