@grendello 2022-04-13
Fix for the failing tests: https://github.com/xamarin/xamarin-android/pull/6922